graph and favourites services renamed. App and env params added to both of them.
